mysite/requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cleanup==6.0.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
psycopg2-binary
psycopg2>=2.8
sqlparse==0.4.2
tzdata==2022.1
